project-atlas/requirements.txt
playwright==1.42.0
beautifulsoup4
python-dotenv
google-genai
openpyxl
pypdf